HCMatcher: Added matches:describingMismatchTo: Added describeMismatchOf:to:
Both methods have default implementations in HCBaseMatcher.
Most implementations can just implement matches: (as before), letting the default implementation of matches:describingMismatchTo: call it. But if it makes more sense to generate the mismatch description during the matching, override matches:describingMismatchTo: and have matches: call it with a nil description.